Backend_api_new/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Werkzeug
itsdangerous
pymysql
flask_mysqldb
flask_cors
flask_session